Add functionality and a drawer for configs in a profile. Fix a bug in LoaderProfile.
Renamed ScriptableObjectDrawer to DisplayDrawerAttributeDrawer and made it an attribute to make it more generic.

Added LoaderConfigTypeAndInstance to make it easily configurable in the inspector. Adds abstraction and ensures there is a type of each instance, rather than 2 lists that assume there is.

Added a drawer for LoaderConfigTypeAndInstanceDrawer to easliy configure.

Fixed the serialisation in LoaderProfile.  Converted it to a Linq statement. Also, it and deserialisation were inverted.

There is a bug with the DisplayDrawerAttributeDrawer where a list does not get hidden if its foldout is inactive, despite the drawer returning the correct height.

The foldout for a list in DisplayDrawerAttributeDrawer seems to only be updated by hiding the outermost foldout, click off the object to destroy the drawer, then click back on the object to recreate it.