Fix LoaderProfile deserialization, LoaderProfile config setting and NullReferenceException when creating the profiles editor.
LoaderProfile now inherits ISerializationCallbackReceiver as intended.
LoaderProfile configs we being set before LoaderProfile had been deserialised, thus they were always being set to default on load.

An event is raised when an object reference in the DisplayDrawerAttributeDrawer PropertyField changes. When this is invoked, the profile's configs get updated. Previously they were being updated whenever any changes were made to the profile SerializeProperty due to the way the callback registration works on the PropertyField.

Setting profile configs is now done via a method so the internal dictionary can be updated.

Fixed a NullReferenceException when loading the LoaderProfilesEditor with a null profile.

Updated the variable "a" in LoaderProfilesEditor.AddProfilePropertyField to an appropriate name.

Removed Awake from LoaderProfile since it is now unessassary.