Place towns sequentially, and fix player town field and town placement excpeption.
Player.Town could not be assigned to as it was an interface. So a second field has been added to accommodate that.

When assigning town tiles, it was inverting the y-axis assignment thus potentially cause IndexOutOfRangeException. This has been fixed and tests updated to reflect this.

Added an IPlayer to hold up for future mocking.

Town placement now works sequentially from TownLoader.

Unsubscribing from an event can cause an enumeration error that needs to be addressed next.