Remove MonoSimulator and townPlacedEvent, and add town placement factories.
Converted townPlacedEvent to a c# event on ITownPlacement.

Deleted MonoSimulator as Zenject can replicate its functionality.

There's a bug in Event where unregistering for updates throws as it is altering the subscribers enumerable during enumeration.
Added a test to EventTests to check for the above behaviour. The fix has not been implemented as it is unclear if this system is even required.

Updated some test namespaces to be better displayed on the test runner.

Added town placement factories to allow for the return to be swapped out via the Installer.

ManualTownPlacement now inherits ITickable to simulate MonoBehaviour behaviour since MonoSimulator has now been removed.