Add town building prefab and fix Interactable.
Removed Interactable as it was problematic to encompass every type of interaction in one class that we easy to use in the inspector.

Updated Interaction to work as a base class MonoBehaviour for each interaction type to allow them to be more explicitly defined.

Updated PlayModeTestHelper to not require a verify button, also made creating it easier and allow for additional, custom buttons.

Updated how building info works. It is now just Building as an MB DTO.

The building info panel has been moved to the info panel since that is where it belongs, also makes unit testing easier.

Since panel controllers should now love on the info panel GO, there is no need for the info panel references hence it has been removed.