Add ability to select GameObjects, update InputTestsBase and remove IStandardInput.
GameObjects can be selected via mouse/touch press and release. If one is selected, it can be unselected by pressing and releasing elsewhere.

IStandardInput will not be mocked since Unity provides the ability to test drive input. The generated Inputs class will receive the input as it would at runtime, functioning the same.