Make UIBuildingInfoPanel a component for the GameObject in question, fix a bug in SelectedObjectTracker and SelectedObjectTrackerTests and add PlayerFactory.
Made UIBuildingInfoPanel a component of the building that wishes to display its info rather then it being on the info panel itself. This is to allow it to work with prefabs since they can't reference the info panel as it is in the scene, but it can be injected into UIBuildingInfoPanel.

Fixed a bug in the SelectedObjectTrackerTests where OnDeselected was not checking the callback correctly. It was evaluating the selectable (implicit bool) rather than comparing it with what the event passed in.

Added PlayerFactory to ensure players are injected into.

Fixed a bug in SelectedObjectTracker that was causing it to always pass null as the deselected object.

Fixed a bug in GameSetupLoader where it was not being injected into.

GameSetupLoader now uses PlayerFactory.

GameFlow no longer uses Awake as it was being called before Construct. Instead, Construct should be used in place of Awake.

TurnManager now uses IPlayer over Player.