Add info to buildings and a IUIContainter.teardown.
Added info to town buildings. Title, default level and description.

IUIContainter.teardown is called when it is either hidden or shown. This is to ensure clean up. For example, unsubscribing from a button click when the panel is hidden or shown by another script, you would not want the callback if the panel is no longer relevant to you.