Add resource collection structure.
IMoveResources request resources to be moved between two IHoldResources.

IHaltResources requests movement of resources to be halted between specified destinations.

The system will gather all halt requests; then all movement requests, resolving along the way.
Resolving a movement request requires checking to see if halt requests for it exist. If so, those requests reduce the available resources in the move request by up to the amount in the halt request. The available reduction for the halt request is also reduced by this amount since it has expended that much of the request.
Once all the halt requests have been resolved against the move request, the remaining resources in the move request will then be moved.