Add ResourceRequests tests.
Fixed a bug in ResourceRequests where it was only getting the order for IMoveResources. See below.

Added GetOrderFor<T> for ResourceCollectionOrder so that it can get the order for a generic, whereas before it was specifically getting the order for IMoveResources regardless.

Made sure a null IPlayer cannot be passed into ResourceRequests as that doesn't make sense.

Fixed a bug in ResourceRequests  that was assuming every Type in the collection order had a registered handler.