Convert GameFlow to LoaderTransitions, load loaders via type rather than instance and fix a bug for LoaderProfiles.Current.
Added LoaderTransitions to facilitate game flow.

Removed GameFlow.

Renamed LoaderAbstract to Loader.

Converted Loader to Iloader.

Changed loading to take a type over an instance. Corollary: all loaders no longer require references for dependencies.

Added LoaderFactory to account for the above.

Fixed a bug in LoaderProfiles where the editor was unable to fetch the "Current" property.

Added a warning message when attempting to load an already loaded Loader.

Added two events to signal when a loader has finished loading: TopLevelLoaded (the be commonly used) and DependencyLoaded (not sure of use).