Fix HUD UI issues, add PlayerHUD.UpdateValues, fix resources order bug and fix UI not being updated correctly for play mode tests.
Updated the HUD UI to fix borders, alignment and scaling.

Added an option in DialogBox to close on option selection, rather than auto closing. Not closing helps with the HUD tests.

Updated PlayerHUD to use ICharacter rather than IPlayer.

Added PlayerHUD.UpdateValues.

Resources are displayed in the enum order rather than the dict keys from a Resources object to ensure a consistent order. This also makes it easy to adjust the order.

Added a colon to the end of a resource name in PlayerHUD.

Added new HUD tests and adjusted existing ones to check for uniqueness, alignment and scaling.

Fixed a bug for play mode tests where the UI wasn't getting updated. Ths was fixed by adding a camera to the scene in PlayModeTestHelper.