Add end turn button and fix a bug in PlayerModeTestHelper.
Created a HUD uxml sheet to compose the different sections and Updated the UIContainer accordingly.

Created the end turn button and added it to the HUD uxml.

Fixed a bug in the PlayModeTestHelper where the test was not concluded after a failed assert in the callback of a custom button.