Begin adding the "Get resource holders" system.
Added ability to add/remove entities to/from a tile and be able to get them based on interface inheritance (IHoldResources for example).

Added AllOnTileGRH which gets resources holders that are on specified tile. Currently just the tile a given IPlaceableInTile is on.

Removed IUpdateable and Updateable. These were mono simulation classes that are not necessary as Zenject has its own.