Add IPlaceable, ITileInfo now uses Resources and Merged Tile and TileMono.
Added IPlaceable for entities that can be placed on the board. Town, unit etc.

ITile inherits IProduceResources.

ITileInfo now uses Resources instead of ResourceType to be consistent with other systems.

Merged Tile and TileMono so there is not longer a need for the separation. Tests that need a tile factory use the new MockTileFactory.

IProduceResources no longer inherits IHandleResources as it required  an IPlayer property, which is uneccasary.