Add DynamicEnumFlags and store selected enum values as string.
Values are now stored as strings in DynamicEnum and DynamicEnumFlags so they are independent of the order defined in DynamicEnumValues.

All instances of the enums will need to be propergated however, to sync their selected values with the user-defined values. This will likely be done by registering each instance of enum to a helper class when they're created, the just looping through them for validation when the user-defined values get updated.

Currently, the only issue with not syncing is the possibility of a value being deleted. This value would still exist on enums that had it selected causing a small amount of wasted memory and possibly an unwanted selection should the value be added back.