Add DynamicEnumFlagsCleanup and make enum values have associated ids.
Renamed EditorUtilities to AssetUtilities.

Converted enum flags to store ids rather than values to allow for faster comparison.

Added a requirement to save changes when editing enum values. Since they now work using an id system, value changes require the ids to be updated and for deleted values to be removed from the ids.

Made the enum values hidden in the inspector so it is required to edit them via the editor window.

Added DynamicEnumFlagsCleanup to remove any ids in DynamicEnumFlags that are no longer bound to a value (because the value was deleted). Need to add a class that loads all scenes so all enums can be cleaned up before a build. It's a small amount of memory wasted through normal use so it doesn't matter that they are not cleaned up in the editor.