Add ability to edit multiple dynamic enum types simultaneously and prompt for outstanding changes.
Added information label to EditDynamicEnumValuesWindow to mention duplicate values and if an error occurs loading values.

The window now supports editing multiple different enums simultaneously with a new window for each type. Each window rect is saved.

Added tracking to unsaved changes and Unity now prompts the user to save/discard when attempting to close the window if there are any outstanding changes.

Renamed the label on the editing values field to make it nicer for the user as they don't need to know what the field is actually called, and the editing is implicit.

Fixed bug in AssetUtilities.TryGetAsset where it would return the first asset using the given name as a substring. It now returns the first with the exact name matching.

Fixed bug in NoAssetsFoundException where it was loging "type" instead of the name of the given type.

Added UnitTypea for testing.

There's a bug where the edited values are not being saved.