Fix UnitMovementSelectorTests, a bug in AssetUtilities search, PointerOverObjectTracker event null bug. Update MouseUtilities.
Added BoxCollider to Tile prefab to allow it to be detected via pointer hover.

Fixed a bug in AssetUtilities.GetAsset that returned the first asset with the given name regardless of type. Added selection parameters that are used in the AssetDatabase search to fix this.

Fixed a bug in PointerOverObjectTracker that didn't null check before invoking events.

Removed unnecessary checking for the pointer over a UI element in PointerOverObjectTracker and SelectedObjectTracker. This is already handled in the CastAllTo3D call they both do.

In UnitMovementSelector, moved pointer hover event (un)subscribing from the constructor to the methods that are invoked from a unit being selected. These events do not need to be listened to unless a unit is selected.

Set the PointerOverObjectTracker mode to TrackMany in UnitMovementSelector to allow tracking of tiles underneath units. This is changed to TrackFirst when a unit is deselected and should be handled more appropriately as it could cause clashes with other system tying to alter the tracking mode.

In UnitMovementSelector, when a GameObejct is hovered/unhovered, the unit slots will show/hide availability.

MouseUtilities can now automatically call PointerOverObjectTracker.FixedTick when the mouse is interacted with. This avoids wet code with tests doing this themselves and possibly having a wrapper method for it.

Added a TODO bug fix for PlayModeTestHelper.Pause.

TestUtilities now uses AssetUtilities to load a prefab rather than it's own, outdated, system.

Fixed some tests, added utility and cleaned up UnitMovementSelectorTests.